created the useLogin hook
